move mpark::variant into include/spdlog so its installed with other headers
