+++ include/spdlog/variant.hpp
